WIP FLANN standalone

+++ src/tools/normals/Main.cpp
@@ -113,7 +113,7 @@ void interpolateNormals(PointBufferPtr pc, size_t numPoints, int n)
     {
         // Create search tree
         vector< ulong > indices;
-        vector< double > distances;
+        vector< float > distances;
 
         Vertex<float> vertex(points[3 * i], points[3 * i + 1], points[3 * i + 2]);
         tree->kSearch( vertex, n, indices, distances);
